feat: exposing raw api implementation
