+++ examples/js/react-native/gnoboard/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
-    "start": "expo start",
+    "start": "expo start -d",
     "android": "npx react-native run-android",
     "ios": "npx react-native run-ios",
     "web": "expo start --web",
