+++ expo/example/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
-    "start": "expo start",
+    "start": "expo start -d",
     "android": "expo run:android",
     "ios": "expo run:ios",
     "web": "expo start --web"
