+++ expo/src/api/index.ts
@@ -1,0 +1,2 @@
+export * from './GnoNativeApi';
+export * from './types';
