--- expo/src/hooks/types.ts
@@ -1,19 +1,0 @@
-import { KeyInfo } from '@buf/gnolang_gnonative.bufbuild_es/gnonativetypes_pb';
-
-export type GnoConfig = {
-  Remote: string;
-  ChainID: string;
-  KeyName: string;
-  Password: string;
-  GasFee: string;
-  GasWanted: number;
-  Mnemonic: string;
-};
-
-export type NetworkMetainfo = {
-  chainId: string;
-  chainName: string;
-  gnoAddress: string;
-};
-
-export type GnoAccount = KeyInfo;
